Refactor after code review
