+++ apps/api/src/app/app.module.ts
@@ -16,7 +16,7 @@ import { FileUploadModule } from '@island.is/api/domains/file-upload'
 import { DocumentModule } from '@island.is/api/domains/documents'
 import { CommunicationsModule } from '@island.is/api/domains/communications'
 import { EmailSignupModule } from '@island.is/api/domains/email-signup'
-import { EmailSignupConfig } from '@island.is/api/domains/email-signup'
+import { ZenterSignupConfig } from '@island.is/api/domains/email-signup'
 import { CmsTranslationsModule } from '@island.is/cms-translations'
 import { UserProfileModule } from '@island.is/api/domains/user-profile'
 import { NationalRegistryModule } from '@island.is/api/domains/national-registry'
@@ -289,7 +289,7 @@ const autoSchemaFile = environment.production
         FileStorageConfig,
         FiskistofaClientConfig,
         PowerBiConfig,
-        EmailSignupConfig,
+        ZenterSignupConfig,
       ],
     }),
   ],
