+++ libs/api/domains/email-signup/src/index.ts
@@ -1,3 +1,3 @@
 export * from './lib/emailSignup.module'
 export * from './lib/emailSignup.resolver'
-export * from './lib/emailSignup.config'
+export * from './lib/services/zenter/zenter.config'
