+++ libs/api/domains/email-signup/src/lib/constants/index.ts
@@ -1,0 +1,5 @@
+export enum FormFieldType {
+  CHECKBOXES = 'checkboxes',
+}
+
+export const ZENTER_IMPORT_ENDPOINT_URL = 'https://samskipti.zenter.is/import'
