+++ libs/api/domains/email-signup/src/lib/emailSignup.module.ts
@@ -1,10 +1,19 @@
 import { Module } from '@nestjs/common'
+
 import { CmsModule } from '@island.is/cms'
+
 import { EmailSignupResolver } from './emailSignup.resolver'
-import { EmailSignupServiceProvider } from './serviceProvider'
+import { EmailSignupService } from './emailSignup.service'
+import { ZenterSignupService } from './services/zenter/zenter.service'
+import { MailchimpSignupService } from './services/mailchimp/mailchimp.service'
 
 @Module({
   imports: [CmsModule],
-  providers: [EmailSignupServiceProvider, EmailSignupResolver],
+  providers: [
+    ZenterSignupService,
+    MailchimpSignupService,
+    EmailSignupService,
+    EmailSignupResolver,
+  ],
 })
 export class EmailSignupModule {}
